Refactor: 요청사항 수정하기 API 수정
